housing-api/requirements.txt
Flask
psycopg2-binary
flask-migrate
flask_sqlalchemy
